tempfile::env

Function override_temp_dir

pub fn override_temp_dir(path: &Path) -> Result<(), PathBuf>

Override the default temporary directory (defaults to std::env::temp_dir). This function changes the global default temporary directory for the entire program and should not be called except in exceptional cases where it’s not configured correctly by the platform. Applications should first check if the path returned by env::temp_dir is acceptable.

If you’re writing a library and want to control where your temporary files are placed, you should instead use the _in variants of the various temporary file/directory constructors (tempdir_in, tempfile_in, the so-named functions on Builder, etc.).

Only the first call to this function will succeed. All further calls will fail with Err(path) where path is previously set default temporary directory override.

NOTE: This function does not check if the specified directory exists and/or is writable.