Struct tempfile::NamedTempFileBuilder

pub struct NamedTempFileBuilder<'a, 'b> { /* fields omitted */ }

Create a new temporary file with custom parameters.

Security

This variant is NOT secure/reliable in the presence of a pathological temporary file cleaner.

Resource Leaking

NamedTempFiles are deleted on drop. As rust doesn't guarantee that a struct will ever be dropped, these temporary files will not be deleted on abort, resource leak, early exit, etc.

Methods

impl<'a, 'b> NamedTempFileBuilder<'a, 'b>

fn new() -> Self

Create a new NamedTempFileBuilder.

Examples

use tempfile::NamedTempFileBuilder;

let named_temp_file = NamedTempFileBuilder::new()
    .prefix("my-temporary-note")
    .suffix(".txt")
    .rand_bytes(5)
    .create()?;
 
let name = named_temp_file
    .path()
    .file_name()?
    .to_str()?;

assert!(name.starts_with("my-temporary-note"));
assert!(name.ends_with(".txt"));
assert_eq!(name.len(), "my-temporary-note.txt".len() + 5);

fn prefix(&mut self, prefix: &'a str) -> &mut Self

Set a custom filename prefix.

Path separators are legal but not advisable. Default: .tmp.

Examples

TODO

fn suffix(&mut self, suffix: &'b str) -> &mut Self

Set a custom filename suffix.

Path separators are legal but not advisable. Default: empty.

Examples

TODO

fn rand_bytes(&mut self, rand: usize) -> &mut Self

Set the number of random bytes.

Default: 6.

Examples

TODO

fn create(&self) -> Result<NamedTempFile>

Create the named temporary file.

Security

See: NamedTempFile::new

Errors

TODO

Examples

TODO

fn create_in<P: AsRef<Path>>(&self, dir: P) -> Result<NamedTempFile>

Create the named temporary file in the specified directory.

Errors

TODO

Examples

TODO