tempfile

Struct SpooledTempFile

pub struct SpooledTempFile { /* private fields */ }

An object that behaves like a regular temporary file, but keeps data in memory until it reaches a configured size, at which point the data is written to a temporary file on disk, and further operations use the file on disk.

Implementations

impl SpooledTempFile

pub fn new(max_size: usize) -> SpooledTempFile

Construct a new SpooledTempFile.

pub fn new_in<P: AsRef<Path>>(max_size: usize, dir: P) -> SpooledTempFile

Construct a new SpooledTempFile, backed by a file in the specified directory.

pub fn is_rolled(&self) -> bool

Returns true if the file has been rolled over to disk.

pub fn roll(&mut self) -> Result<()>

Rolls over to a file on disk, regardless of current size. Does nothing if already rolled over.

pub fn set_len(&mut self, size: u64) -> Result<(), Error>

Truncate the file to the specified size.

pub fn into_inner(self) -> SpooledData

Consumes and returns the inner SpooledData type.

pub fn into_file(self) -> Result<File>

Convert into a regular unnamed temporary file, writing it to disk if necessary.