tempfile::env
Function temp_dir

pub fn temp_dir() -> PathBuf

Returns the default temporary directory, used for both temporary directories and files if no directory is explicitly specified.

This function simply delegates to std::env::temp_dir unless the default temporary directory has been override by a call to override_temp_dir.

NOTE: This function does check if the returned directory exists and/or is writable.