Module tempfile::file

Modules

imp

Structs

NamedTempFile

A named temporary file.

PathPersistError

Error returned when persisting a temporary file path fails.

PersistError

Error returned when persisting a temporary file fails.

TempPath

A path to a named temporary file without an open file handle.

Functions

create_named

tempfile

Create a new temporary file.

tempfile_in

Create a new temporary file in the specified directory.