Function tempfile::tempdir_in

pub fn tempdir_in<P: AsRef<Path>>(dir: P) -> Result<TempDir>

Create a new temporary directory in a specific directory.

The tempdir_in function creates a directory in the specified directory and returns a TempDir. The directory will be automatically deleted when the TempDirs destructor is run.

Resource Leaking

See the resource leaking docs on TempDir.

Errors

If the directory can not be created, Err is returned.

Examples

use tempfile::tempdir_in;
use std::fs::File;
use std::io::{self, Write};

// Create a directory inside of the current directory.
let tmp_dir = tempdir_in(".")?;

let file_path = tmp_dir.path().join("my-temporary-note.txt");
let mut tmp_file = File::create(file_path)?;
writeln!(tmp_file, "Brian was here. Briefly.")?;

// `tmp_dir` goes out of scope, the directory as well as
// `tmp_file` will be deleted here.
drop(tmp_file);
tmp_dir.close()?;