Struct tempfile::NamedTempFile

pub struct NamedTempFile { /* fields omitted */ }

A named temporary file.

The default constructor, NamedTempFile::new(), creates files in the location returned by std::env::temp_dir(), but NamedTempFile can be configured to manage a temporary file in any location by constructing with NamedTempFile::new_in().

Security

This variant is NOT secure/reliable in the presence of a pathological temporary file cleaner.

Resource Leaking

If the program exits before the NamedTempFile destructor is run, such as via std::process::exit(), by segfaulting, or by receiving a signal like SIGINT, then the temporary file will not be deleted.

Use the tempfile() function unless you absolutely need a named file.

Methods

impl NamedTempFile

pub fn new() -> Result<NamedTempFile>

Create a new named temporary file.

See Builder for more configuration.

Security

This will create a temporary file in the default temporary file directory (platform dependent). These directories are often patrolled by temporary file cleaners so only use this method if you're positive that the temporary file cleaner won't delete your file.

Reasons to use this method:

The file has a short lifetime and your temporary file cleaner is sane (doesn't delete recently accessed files).

You trust every user on your system (i.e. you are the only user).

You have disabled your system's temporary file cleaner or verified that your system doesn't have a temporary file cleaner.

Reasons not to use this method:

You'll fix it later. No you won't.

You don't care about the security of the temporary file. If none of the "reasons to use this method" apply, referring to a temporary file by name may allow an attacker to create/overwrite your non-temporary files. There are exceptions but if you don't already know them, don't use this method.

Errors

If the file can not be created, Err is returned.

Examples

Create a named temporary file and write some data to it:

use tempfile::NamedTempFile;

let mut file = NamedTempFile::new()?;

writeln!(file, "Brian was here. Briefly.")?;

pub fn new_in<P: AsRef<Path>>(dir: P) -> Result<NamedTempFile>

Create a new named temporary file in the specified directory.

See NamedTempFile::new() for details.

pub fn path(&self) -> &Path

Get the temporary file's path.

Security

Only use this method if you're positive that a temporary file cleaner won't have deleted your file. Otherwise, the path returned by this method may refer to an attacker controlled file.

Examples

use tempfile::NamedTempFile;

let file = NamedTempFile::new()?;

println!("{:?}", file.path());

pub fn close(self) -> Result<()>

Close and remove the temporary file.

Use this if you want to detect errors in deleting the file.

Errors

If the file cannot be deleted, Err is returned.

Examples

use tempfile::NamedTempFile;

let file = NamedTempFile::new()?;

// By closing the `NamedTempFile` explicitly, we can check that it has
// been deleted successfully. If we don't close it explicitly,
// the file will still be deleted when `file` goes out
// of scope, but we won't know whether deleting the file
// succeeded.
file.close()?;

pub fn persist<P: AsRef<Path>>(self, new_path: P) -> Result<File, PersistError>

Persist the temporary file at the target path.

If a file exists at the target path, persist will atomically replace it. If this method fails, it will return self in the resulting PersistError.

Note: Temporary files cannot be persisted across filesystems.

Security

Only use this method if you're positive that a temporary file cleaner won't have deleted your file. Otherwise, you might end up persisting an attacker controlled file.

Errors

If the file cannot be moved to the new location, Err is returned.

Examples

use tempfile::NamedTempFile;

let file = NamedTempFile::new()?;

let mut persisted_file = file.persist("./saved_file.txt")?;
writeln!(persisted_file, "Brian was here. Briefly.")?;

pub fn persist_noclobber<P: AsRef<Path>>(
self,
new_path: P
) -> Result<File, PersistError>

Persist the temporary file at the target path iff no file exists there.

If a file exists at the target path, fail. If this method fails, it will return self in the resulting PersistError.

Note: Temporary files cannot be persisted across filesystems. Also Note: This method is not atomic. It can leave the original link to the temporary file behind.

Security

Only use this method if you're positive that a temporary file cleaner won't have deleted your file. Otherwise, you might end up persisting an attacker controlled file.

Errors

If the file cannot be moved to the new location or a file already exists there, Err is returned.

Examples

use tempfile::NamedTempFile;

let file = NamedTempFile::new()?;

let mut persisted_file = file.persist_noclobber("./saved_file.txt")?;
writeln!(persisted_file, "Brian was here. Briefly.")?;

pub fn reopen(&self) -> Result<File>

Reopen the temporary file.

This function is useful when you need multiple independent handles to the same file. It's perfectly fine to drop the original NamedTempFile while holding on to Files returned by this function; the Files will remain usable. However, they may not be nameable.

Errors

If the file cannot be reopened, Err is returned.

Examples

use tempfile::NamedTempFile;

let file = NamedTempFile::new()?;

let another_handle = file.reopen()?;

pub fn as_file(&self) -> &File

Get a reference to the underlying file.

pub fn as_file_mut(&mut self) -> &mut File

Get a mutable reference to the underlying file.

pub fn into_file(self) -> File

Convert the temporary file into a std::fs::File.

The inner file will be deleted.

pub fn into_temp_path(self) -> TempPath

Closes the file, leaving only the temporary file path.

This is useful when another process must be able to open the temporary file.