Struct tempfile::SpooledTempFile

pub struct SpooledTempFile { /* fields omitted */ }

An object that behaves like a regular temporary file, but keeps data in memory until it reaches a configured size, at which point the data is written to a temporary file on disk, and further operations use the file on disk.

Implementations

impl SpooledTempFile

pub fn new(max_size: usize) -> SpooledTempFile

pub fn is_rolled(&self) -> bool

Returns true if the file has been rolled over to disk.

pub fn roll(&mut self) -> Result<()>

Rolls over to a file on disk, regardless of current size. Does nothing if already rolled over.

pub fn set_len(&mut self, size: u64) -> Result<(), Error>