Struct tempfile::TempFile

pub struct TempFile(_);

An unnamed temporary file.

This variant is secure/reliable in the presence of a pathological temporary file cleaner.

Deletion:

Linux >= 3.11: The temporary file is never linked into the filesystem so it can't be leaked.

Other *nix: The temporary file is immediately unlinked on create. The OS will delete it when the last open copy of it is closed (the last TempFile reference to it is dropped).

Windows: The temporary file is marked DeleteOnClose and, again, will be deleted when the last open copy of it is closed. Unlike *nix operating systems, the file is not immediately unlinked from the filesystem.

Methods

impl TempFile

pub fn new() -> Result<TempFile>

Create a new temporary file.

pub fn new_in<P: AsRef<Path>>(dir: P) -> Result<TempFile>

Create a new temporary file in the specified directory.

pub fn shared(count: usize) -> Result<Vec<TempFile>>

Create a new temporary file and open it count times returning count independent references to the same file.

This can be useful if you want multiple seek positions in the same temporary file. Additionally, this function guarantees that all of the returned temporary file objects refer to the same underlying temporary file even in the presence of a pathological temporary file cleaner.

pub fn shared_in<P: AsRef<Path>>(dir: P, count: usize) -> Result<Vec<TempFile>>

Same as shared but creates the file in the specified directory.

pub fn len(&self) -> Result<u64>

Number of bytes in the file.

pub fn set_len(&self, size: u64) -> Result<()>

Truncate the file to size bytes.